Add chain of responsibility and null object patterns
There are still bugs in player movement

+++ src/patterns/factories/enemies/Enemy.java
@@ -23,11 +23,11 @@ public class Enemy extends Unit {
     }
 
     @Override
-    public void moveX(int value) {
+    public void moveX(double value) {
     }
 
     @Override
-    public void moveY(int value) {
+    public void moveY(double value) {
     }
 
     @Override
